requirements.txt
Django==4.1.2
django-filter==22.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
sqlparse==0.4.3
django-cors-headers==3.13.0
django-environ==0.9.0
environ
requests
djangorestframework-simplejwt
djangorestframework-jwt